Add files via upload
This website is built using html and css. It allows the user to choose a music genre by clicking on the image of the genre that directs the user to a page that includes the artists of that genre and on clicking on the image of the artist it directs the user to the official website of the artist and on clicking on details the user can choose to view the spotify , instagram accounts and wikipedia of that particular artist.